api-types, ch-remote, fuzz, vmm: split `VmReceiveMigrationData`
Part of the move to an api-types crate.

On-behalf-of: SAP julian.schindel@sap.com
