+++ vmm/src/api/mod.rs
@@ -33,7 +33,6 @@
 pub mod dbus;
 pub mod http;
 
-use std::collections::HashSet;
 use std::io;
 use std::num::{NonZeroU32, NonZeroU64};
 use std::path::PathBuf;
@@ -43,27 +42,28 @@ use std::time::Duration;
 #[cfg(all(target_arch = "x86_64", feature = "guest_debug"))]
 use api_types::VmCoredumpData;
 use api_types::{
-    MigrationMode, RestoredVfioConfig, TimeoutStrategy, VmMemoryZoneUpdateData, VmRemoveDeviceData,
-    VmResizeData, VmResizeDiskData, VmResizeZoneData, VmSnapshotConfig, VmmPingResponse,
+    MigrationMode, TimeoutStrategy, VmRemoveDeviceData, VmResizeData, VmResizeDiskData,
+    VmResizeZoneData, VmSnapshotConfig, VmmPingResponse,
 };
 use log::info;
 use micro_http::Body;
-use option_parser::{OptionParser, OptionParserError, Toggle, Tuple, TupleList};
+use option_parser::{OptionParser, OptionParserError, Toggle};
 use serde::{Deserialize, Serialize};
 use thiserror::Error;
 use vm_migration::MigratableError;
-use vm_migration::tls::{TlsConfigError, TlsEndpoint, validate_tls_dir};
+use vm_migration::tls::{TlsEndpoint, validate_tls_dir};
 use vmm_sys_util::eventfd::EventFd;
 
 #[cfg(feature = "dbus_api")]
 pub use self::dbus::start_dbus_thread;
 pub use self::http::{start_http_fd_thread, start_http_path_thread};
 use crate::Error as VmmError;
-use crate::config::{RestoreConfig, deserialize_restored_fd};
+use crate::config::RestoreConfig;
 use crate::device_tree::DeviceTree;
 use crate::migration::transport::{
     MAX_MIGRATION_CONNECTIONS, TcpAddressParseError, tcp_address_to_server_name,
 };
+use crate::migration::{VmReceiveMigrationConfigError, VmReceiveMigrationData};
 use crate::vm::{Error as VmError, VmState};
 use crate::vm_config::{
     DeviceConfig, DiskConfig, FsConfig, GenericVhostUserConfig, NetConfig, PmemConfig,
@@ -221,6 +221,10 @@ pub enum ApiError {
     /// Error triggering NMI
     #[error("Error triggering NMI")]
     VmNmi(#[source] VmError),
+
+    /// The receive migration configuration is invalid.
+    #[error("The receive migration configuration is invalid")]
+    VmReceiveMigrationConfig(#[source] VmReceiveMigrationConfigError),
 }
 pub type ApiResult<T> = Result<T, ApiError>;
 
@@ -231,250 +235,6 @@ pub struct VmInfoResponse {
     pub state: VmState,
     pub memory_actual_size: u64,
     pub device_tree: Option<DeviceTree>,
-}
-
-#[derive(Clone, Deserialize, Serialize, Default, Debug)]
-#[cfg_attr(test, derive(PartialEq))]
-pub struct VmReceiveMigrationData {
-    /// URL for the reception of migration state
-    pub receiver_url: String,
-    /// Directory containing the TLS server certificate (`server-cert.pem`),
-    /// the TLS server key (`server-key.pem`), and the server's TLS root CA
-    /// certificate (`ca-cert.pem`).
-    ///
-    /// If this is `Some`, the migration is instructed to use mTLS.
-    #[serde(default)]
-    pub tls_dir: Option<PathBuf>,
-    /// Memory transfer mode.
-    #[serde(default)]
-    pub memory_mode: MigrationMode,
-    /// Optional VFIO device id to cdev FD pairs, used to substitute each
-    /// device's saved path or stale FD in the received VmConfig.
-    #[serde(default)]
-    pub vfio_fds: Option<Vec<RestoredVfioConfig>>,
-    // FDs are not serialized and any deserialized value is invalid; see NetConfig::fds.
-    #[serde(default, deserialize_with = "deserialize_restored_fd")]
-    pub iommufd_fd: Option<i32>,
-    /// Optional memory zone update data
-    #[serde(default)]
-    pub zone_updates: Vec<VmMemoryZoneUpdateData>,
-}
-
-#[derive(Debug, Error)]
-pub enum VmReceiveMigrationConfigError {
-    /// Parsing of `VmReceiveMigrationData` failed
-    #[error("Error parsing receive migration parameters")]
-    ParseError(#[source] OptionParserError),
-    /// Variant returned of validation of `receiver_url` failed.
-    #[error("Expected receiver_url in the form of either `tcp:<host>:<port>` or `unix:<path>:`")]
-    MalformedReceiverUrl(#[source] TcpAddressParseError),
-    /// TLS encryption cannot be used for UNIX sockets. It is therefore
-    /// forbidden to use TLS encryption with UNIX sockets.
-    #[error("UNIX sockets and TLS encryption cannot be used at the same time")]
-    TlsEncryptionUsedForUnixSocket,
-    /// The `receiver_url` does not contain one of the supported
-    /// prefixes. Supported prefixes are "tcp" and "unix".
-    #[error("Expected receiver_url to either use `tcp` or `unix` prefix")]
-    InvalidSocketPrefix,
-    /// The TLS configuration is invalid.
-    #[error("Invalid TLS configuration for receive-migration")]
-    InvalidTlsConfiguration(#[source] TlsConfigError),
-    /// Every VFIO device needs a replacement in vfio_fds and none was found for the respective
-    /// device
-    #[error(
-        "VFIO device '{0}' has no replacement in vfio_fds, its source path or fd is not usable on the destination"
-    )]
-    VfioDeviceNoReplacementFd(String),
-    /// The `vfio_fds` option was used without supplying `iommufd_fd`
-    #[error("Usage of `vfio_fds` requires `iommufd_fd` to be specified")]
-    VfioFdRequiresIommufdFd,
-    /// `iommufd_fd` was provided without also enabling the iommufd backend.
-    #[error("Platform `iommufd_fd=<fd>` requires `iommufd=on`")]
-    IommufdFdRequiresIommufd,
-    /// Identified duplicate fd replacements for the FD with the given ID.
-    #[error("Multiple replacements defined for fd with ID {0}")]
-    VfioFdMultipleReplacements(String),
-    /// Identified a fd replacements for a FD not present in the VmConfig.
-    #[error("Replacement ID {0} in vfio_fds id does not match any device in the received VmConfig")]
-    VfioFdReplacementWithoutTarget(String),
-    /// Multiple updates for the same memory zone were defined.
-    #[error("More than one update was defined for at least one memory zone")]
-    MultipleMemoryZoneUpdates,
-    /// An update with an empty memory zone ID was specified.
-    #[error("One or more memory zone updates contained an empty ID")]
-    MemoryZoneUpdatesEmptyId,
-}
-
-impl VmReceiveMigrationData {
-    pub const SYNTAX: &'static str = "VM receive migration parameters \
-        \"<receiver_url>\" or \"receiver_url=<url>[,tls_dir=<path>][,memory_mode=precopy|postcopy]\
-        [,vfio_fds=<list_of_vfio_ids_with_their_associated_fd>][,iommufd_fd=<fd>]\
-        [,zone_updates=[<id@host_numa_node>]]\"";
-
-    pub fn parse(migration: &str) -> Result<Self, VmReceiveMigrationConfigError> {
-        let mut parser = OptionParser::new();
-        parser
-            .add("receiver_url")
-            .add("tls_dir")
-            .add("memory_mode")
-            .add("vfio_fds")
-            .add("iommufd_fd")
-            .add("zone_updates");
-        parser
-            .parse(migration)
-            .map_err(VmReceiveMigrationConfigError::ParseError)?;
-
-        let receiver_url = parser.get("receiver_url").ok_or_else(|| {
-            VmReceiveMigrationConfigError::ParseError(OptionParserError::InvalidSyntax(
-                "receiver_url is required".to_string(),
-            ))
-        })?;
-        let tls_dir = parser
-            .convert::<String>("tls_dir")
-            .map_err(VmReceiveMigrationConfigError::ParseError)?
-            .map(|path| PathBuf::from(&path));
-        let memory_mode = parser
-            .convert::<MigrationMode>("memory_mode")
-            .map_err(VmReceiveMigrationConfigError::ParseError)?
-            .unwrap_or_default();
-        let vfio_fds = parser
-            .convert::<TupleList<String, u64>>("vfio_fds")
-            .map_err(VmReceiveMigrationConfigError::ParseError)?
-            .map(|v| {
-                v.0.iter()
-                    .map(|Tuple(id, fd)| RestoredVfioConfig {
-                        id: id.clone(),
-                        fd: Some(*fd as i32),
-                    })
-                    .collect()
-            });
-        let iommufd_fd = parser
-            .convert::<i32>("iommufd_fd")
-            .map_err(VmReceiveMigrationConfigError::ParseError)?;
-
-        let zone_updates: Vec<VmMemoryZoneUpdateData> = parser
-            .convert::<TupleList<String, u32>>("zone_updates")
-            .map_err(VmReceiveMigrationConfigError::ParseError)?
-            .map_or(Vec::new(), |v| {
-                v.0.iter()
-                    .map(|Tuple(id, host_numa_node)| VmMemoryZoneUpdateData {
-                        id: id.clone(),
-                        host_numa_node: *host_numa_node,
-                    })
-                    .collect()
-            });
-
-        let data = Self {
-            receiver_url,
-            tls_dir,
-            memory_mode,
-            vfio_fds,
-            iommufd_fd,
-            zone_updates,
-        };
-
-        data.validate()?;
-
-        Ok(data)
-    }
-
-    pub fn validate(&self) -> Result<(), VmReceiveMigrationConfigError> {
-        if let Some(addr) = self.receiver_url.strip_prefix("tcp:") {
-            tcp_address_to_server_name(addr)
-                .map_err(VmReceiveMigrationConfigError::MalformedReceiverUrl)?;
-        } else if self
-            .receiver_url
-            .strip_prefix("unix:")
-            .is_some_and(|path| !path.is_empty())
-        {
-            if self.tls_dir.is_some() {
-                return Err(VmReceiveMigrationConfigError::TlsEncryptionUsedForUnixSocket);
-            }
-        } else {
-            return Err(VmReceiveMigrationConfigError::InvalidSocketPrefix);
-        }
-
-        if let Some(tls_dir) = &self.tls_dir {
-            validate_tls_dir(tls_dir, TlsEndpoint::Server)
-                .map_err(VmReceiveMigrationConfigError::InvalidTlsConfiguration)?;
-        }
-
-        let unique_zones = self
-            .zone_updates
-            .iter()
-            .map(|update| update.id.as_str())
-            .collect::<HashSet<_>>();
-        if self.zone_updates.len() != unique_zones.len() {
-            return Err(VmReceiveMigrationConfigError::MultipleMemoryZoneUpdates);
-        }
-        if unique_zones.contains("") {
-            return Err(VmReceiveMigrationConfigError::MemoryZoneUpdatesEmptyId);
-        }
-
-        Ok(())
-    }
-
-    pub fn validate_vfio_fds(
-        &self,
-        vm_config: &VmConfig,
-    ) -> Result<(), VmReceiveMigrationConfigError> {
-        let vfio_fds = self.vfio_fds.as_deref().unwrap_or_default();
-
-        // A migrated VFIO device cannot reuse its source handle. Its fd is
-        // invalid across the migration and its path names the source host's
-        // topology, so every device needs a replacement in vfio_fds. This
-        // holds even when no vfio_fds are supplied at all.
-        let substituted: HashSet<&str> = vfio_fds.iter().map(|v| v.id.as_str()).collect();
-        for d in vm_config.devices.iter().flatten() {
-            if !d
-                .pci_common
-                .id
-                .as_deref()
-                .is_some_and(|id| substituted.contains(id))
-            {
-                return Err(VmReceiveMigrationConfigError::VfioDeviceNoReplacementFd(
-                    d.pci_common.id.as_deref().unwrap_or_default().to_owned(),
-                ));
-            }
-        }
-
-        if vfio_fds.is_empty() {
-            return Ok(());
-        }
-
-        // The supplied vfio_fds must be usable against the received VmConfig.
-        if self.iommufd_fd.is_none() {
-            return Err(VmReceiveMigrationConfigError::VfioFdRequiresIommufdFd);
-        }
-        if !vm_config.platform.as_ref().is_some_and(|p| p.iommufd) {
-            return Err(VmReceiveMigrationConfigError::IommufdFdRequiresIommufd);
-        }
-
-        let mut seen = HashSet::new();
-        for v in vfio_fds {
-            if !seen.insert(v.id.as_str()) {
-                return Err(VmReceiveMigrationConfigError::VfioFdMultipleReplacements(
-                    v.id.to_owned(),
-                ));
-            }
-        }
-
-        let known_ids: HashSet<&str> = vm_config
-            .devices
-            .iter()
-            .flatten()
-            .filter_map(|d| d.pci_common.id.as_deref())
-            .collect();
-        for v in vfio_fds {
-            if !known_ids.contains(v.id.as_str()) {
-                return Err(
-                    VmReceiveMigrationConfigError::VfioFdReplacementWithoutTarget(v.id.to_owned()),
-                );
-            }
-        }
-
-        Ok(())
-    }
 }
 
 #[derive(Debug, Error)]
@@ -1535,16 +1295,20 @@ impl ApiAction for VmReboot {
 pub struct VmReceiveMigration;
 
 impl ApiAction for VmReceiveMigration {
-    type RequestBody = VmReceiveMigrationData;
+    type RequestBody = api_types::VmReceiveMigrationData;
     type ResponseBody = Option<Body>;
 
     fn request(&self, data: Self::RequestBody, response_sender: Sender<ApiResponse>) -> ApiRequest {
         Box::new(move |vmm| {
             info!("API request event: VmReceiveMigration {data:?}");
 
-            let response = vmm
-                .vm_receive_migration(data)
-                .map_err(ApiError::VmReceiveMigration)
+            let response = data
+                .try_into()
+                .map_err(ApiError::VmReceiveMigrationConfig)
+                .and_then(|data: VmReceiveMigrationData| {
+                    vmm.vm_receive_migration(data)
+                        .map_err(ApiError::VmReceiveMigration)
+                })
                 .map(|_| ApiResponsePayload::Empty);
 
             response_sender
@@ -2020,12 +1784,6 @@ mod unit_tests {
             fs::write(self.path.join(file_name), b"test").unwrap();
         }
 
-        fn add_receive_tls_files(&self) {
-            self.add_file("ca-cert.pem");
-            self.add_file("server-cert.pem");
-            self.add_file("server-key.pem");
-        }
-
         fn add_send_tls_files(&self) {
             self.add_file("ca-cert.pem");
             self.add_file("client-cert.pem");
@@ -2037,191 +1795,6 @@ mod unit_tests {
         fn drop(&mut self) {
             let _ = fs::remove_dir_all(&self.path);
         }
-    }
-
-    #[test]
-    fn test_vm_receive_migration_data_parse() {
-        let data = VmReceiveMigrationData::parse("receiver_url=tcp:192.168.1.1:8080").unwrap();
-        assert_eq!(
-            data,
-            VmReceiveMigrationData {
-                receiver_url: "tcp:192.168.1.1:8080".to_string(),
-                tls_dir: None,
-                memory_mode: MigrationMode::Precopy,
-                vfio_fds: None,
-                iommufd_fd: None,
-                zone_updates: vec![],
-            }
-        );
-
-        let data = VmReceiveMigrationData::parse("receiver_url=tcp:[2001:db8::1]:8080").unwrap();
-        assert_eq!(data.receiver_url, "tcp:[2001:db8::1]:8080");
-
-        let data =
-            VmReceiveMigrationData::parse("receiver_url=tcp:destination.example:8080").unwrap();
-        assert_eq!(data.receiver_url, "tcp:destination.example:8080");
-
-        let data = VmReceiveMigrationData::parse("receiver_url=unix:/tmp/ch=migrate.sock").unwrap();
-        assert_eq!(data.receiver_url, "unix:/tmp/ch=migrate.sock");
-
-        let tls_dir = TestDir::new("receive-tls");
-        tls_dir.add_receive_tls_files();
-        let tls_dir_path = tls_dir.path.clone();
-        let data = VmReceiveMigrationData::parse(&format!(
-            "receiver_url=tcp:192.168.1.1:8080,tls_dir={}",
-            tls_dir_path.display()
-        ))
-        .unwrap();
-        assert_eq!(
-            data,
-            VmReceiveMigrationData {
-                receiver_url: "tcp:192.168.1.1:8080".to_string(),
-                tls_dir: Some(tls_dir_path),
-                memory_mode: MigrationMode::Precopy,
-                vfio_fds: None,
-                iommufd_fd: None,
-                zone_updates: vec![],
-            }
-        );
-
-        let tls_dir = TestDir::new("receive-empty-tls");
-        VmReceiveMigrationData::parse(&format!(
-            "receiver_url=tcp:192.168.1.1:8080,tls_dir={}",
-            tls_dir.path.display()
-        ))
-        .unwrap_err();
-
-        let e = VmReceiveMigrationData::parse("receiver_url=file:///tmp/migration").unwrap_err();
-        assert!(
-            matches!(e, VmReceiveMigrationConfigError::InvalidSocketPrefix),
-            "Expected \"{:?}\"; got \"{e:?}\"",
-            VmReceiveMigrationConfigError::InvalidSocketPrefix,
-        );
-        let e = VmReceiveMigrationData::parse("receiver_url=tcp:192.168.1.1").unwrap_err();
-        assert!(
-            matches!(
-                e,
-                VmReceiveMigrationConfigError::MalformedReceiverUrl(
-                    TcpAddressParseError::MissingPort
-                )
-            ),
-            "Expected \"{:?}\"; got \"{e:?}\"",
-            VmReceiveMigrationConfigError::MalformedReceiverUrl(TcpAddressParseError::MissingPort),
-        );
-        let e = VmReceiveMigrationData::parse("receiver_url=tcp:[2001:db8::1]").unwrap_err();
-        assert!(
-            matches!(
-                e,
-                VmReceiveMigrationConfigError::MalformedReceiverUrl(
-                    TcpAddressParseError::MissingPortSeparatorAfterBracketedHost
-                )
-            ),
-            "Expected \"{:?}\"; got \"{e:?}\"",
-            VmReceiveMigrationConfigError::MalformedReceiverUrl(
-                TcpAddressParseError::MissingPortSeparatorAfterBracketedHost
-            ),
-        );
-        let e =
-            VmReceiveMigrationData::parse("receiver_url=unix:/tmp/sock,tls_dir=/tmp").unwrap_err();
-        assert!(
-            matches!(
-                e,
-                VmReceiveMigrationConfigError::TlsEncryptionUsedForUnixSocket
-            ),
-            "Expected \"{:?}\"; got \"{e:?}\"",
-            VmReceiveMigrationConfigError::TlsEncryptionUsedForUnixSocket,
-        );
-
-        // memory_mode defaults to precopy when not specified.
-        let data = VmReceiveMigrationData::parse("receiver_url=tcp:127.0.0.1:1234").unwrap();
-        assert_eq!(
-            data,
-            VmReceiveMigrationData {
-                receiver_url: "tcp:127.0.0.1:1234".to_string(),
-                tls_dir: None,
-                memory_mode: MigrationMode::Precopy,
-                vfio_fds: None,
-                iommufd_fd: None,
-                ..Default::default()
-            }
-        );
-
-        // Explicit receiver_url with memory_mode=postcopy.
-        let data =
-            VmReceiveMigrationData::parse("receiver_url=unix:/tmp/sock,memory_mode=postcopy")
-                .unwrap();
-        assert_eq!(
-            data,
-            VmReceiveMigrationData {
-                receiver_url: "unix:/tmp/sock".to_string(),
-                tls_dir: None,
-                memory_mode: MigrationMode::Postcopy,
-                vfio_fds: None,
-                iommufd_fd: None,
-                ..Default::default()
-            }
-        );
-
-        // Missing receiver_url in keyed form must fail.
-        let e = VmReceiveMigrationData::parse("memory_mode=postcopy").unwrap_err();
-        assert!(
-            matches!(e, VmReceiveMigrationConfigError::ParseError(_)),
-            "Expected \"ParseError\"; got \"{e:?}\"",
-        );
-
-        // vfio_fds without iommufd_fd parses fine now, the pairing is checked
-        // later against the received VmConfig by validate_vfio_fds.
-        let data =
-            VmReceiveMigrationData::parse("receiver_url=tcp:127.0.0.1:1234,vfio_fds=[vfio0@5]")
-                .unwrap();
-        assert!(data.iommufd_fd.is_none());
-
-        // vfio_fds entries with the iommufd FD.
-        let data = VmReceiveMigrationData::parse(
-            "receiver_url=tcp:127.0.0.1:1234,vfio_fds=[vfio0@5,vfio1@7],iommufd_fd=9",
-        )
-        .unwrap();
-        let fds = data.vfio_fds.expect("vfio_fds populated");
-        assert_eq!(fds.len(), 2);
-        assert_eq!(fds[0].id, "vfio0");
-        assert_eq!(fds[0].fd, Some(5));
-        assert_eq!(fds[1].id, "vfio1");
-        assert_eq!(fds[1].fd, Some(7));
-        assert_eq!(data.iommufd_fd, Some(9));
-
-        // zone update tests
-        let e = VmReceiveMigrationData::parse("receiver_url=unix:/tmp/sock,zone_updates=[]")
-            .unwrap_err();
-        assert!(
-            matches!(e, VmReceiveMigrationConfigError::ParseError(_)),
-            "Expected \"ParseError\"; got \"{e:?}\"",
-        );
-        let e = VmReceiveMigrationData::parse("receiver_url=unix:/tmp/sock,zone_updates=[zone1 3]")
-            .unwrap_err();
-        assert!(
-            matches!(e, VmReceiveMigrationConfigError::ParseError(_)),
-            "Expected \"ParseError\"; got \"{e:?}\"",
-        );
-        let e = VmReceiveMigrationData::parse(
-            "receiver_url=unix:/tmp/sock,zone_updates=[zone1@1,zone1@2]",
-        )
-        .unwrap_err();
-        assert!(
-            matches!(e, VmReceiveMigrationConfigError::MultipleMemoryZoneUpdates),
-            "Expected \"{:?}\"; got \"{e:?}\"",
-            VmReceiveMigrationConfigError::MultipleMemoryZoneUpdates,
-        );
-        //Mind the space before the second zone. If the whitespace isn't trimmed, we end up with two
-        //different ID and the error does not trigger
-        let e = VmReceiveMigrationData::parse(
-            "receiver_url=unix:/tmp/sock,zone_updates=[zone1@1, zone1@2]",
-        )
-        .unwrap_err();
-        assert!(
-            matches!(e, VmReceiveMigrationConfigError::MultipleMemoryZoneUpdates),
-            "Expected \"{:?}\"; got \"{e:?}\"",
-            VmReceiveMigrationConfigError::MultipleMemoryZoneUpdates,
-        );
     }
 
     #[test]
